@@ -1,18 +1,18 @@
 {
   "name": "esp-js-ui-module-based-app",
-  "version": "7.0.2",
+  "version": "8.0.14-next.0",
   "license": "MIT",
   "main": "src/app.js",
   "private": true,
   "dependencies": {
     "classnames": "^2.2.6",
-    "esp-js": "^7.0.2",
-    "esp-js-di": "^7.0.2",
-    "esp-js-polimer": "^7.0.2",
-    "esp-js-react": "^7.0.2",
-    "esp-js-ui": "^7.0.2",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
+    "esp-js": "^8.0.14-next.0",
+    "esp-js-di": "^8.0.14-next.0",
+    "esp-js-polimer": "^8.0.14-next.0",
+    "esp-js-react": "^8.0.14-next.0",
+    "esp-js-ui": "^8.0.14-next.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "rxjs": "^6.6.6",
     "uuid": "^8.3.2"
   },
