@@ -15,5 +15,8 @@ export {
 } from './eventTransformations';
 export {EventEnvelopePredicate} from './eventEnvelopePredicate';
 export {PolimerEvents} from './polimerEvents';
-export {sendUpdateToDevTools, connectDevTools} from './reduxDevToolsConnector';
 export {StateHandlerModel} from './stateHandlerModel';
+export {
+    StrictMode,
+    StrictModeSettings,
+} from './strictMode';
