@@ -1,5 +1,5 @@
 import {PolimerEventHandler} from './stateEventHandlers';
-import {connectDevTools, sendUpdateToDevTools} from './reduxDevToolsConnector';
+import {connectReduxDevTools, sendUpdateToReduxDevTools, reduxDevToolsDetectedAndEnabledInEsp} from 'esp-js';
 import {CompositeDisposable, DefaultModelAddress, DisposableBase, EspDecoratorUtil, EventEnvelope, EventObservationMetadata, Guard, Level, ObservationStage, observeEvent, PolimerEventPredicate, Router} from 'esp-js';
 import {EventTransformConfiguration, InputEvent, OutputEvent} from './eventTransformations';
 import {logger} from './logger';
@@ -14,6 +14,24 @@ import {Disposable, utils} from 'esp-js';
 import {StateHandlerConfiguration} from './stateEventHandlers';
 import {StateReaderWriter, MapReaderWriter, DirectStateReaderWriter} from './stateReaderWriter';
 import {EventEnvelopePredicate} from './eventEnvelopePredicate';
+import {createImmutableModelUtility, ImmutableModelUtility} from './immutableModelUtility';
+
+export type DevToolsStateSelector<TModel extends ImmutableModel> = (model: TModel) => object;
+
+export interface DevToolsConfig<TModel extends ImmutableModel> {
+    /**
+     * A list of events which won't be forwarded to dev tools.
+     * This is useful for noisy events which ultimately slow dev tools.
+     */
+    ignoredEvents: string[];
+    /**
+     * A mapper to create state to send to dev tools.
+     * For smaller models you can return the given model sent to the mapper,
+     * for larger models you need to map to a smaller structure to void killing dev tools.
+     * You can also return 'null/undefined' to send no sate (just the event will be sent).
+     */
+    devToolsStateSelector: DevToolsStateSelector<TModel>;
+}
 
 export interface PolimerModelConfig<TModel extends ImmutableModel> {
     initialModel: TModel;
@@ -23,6 +41,7 @@ export interface PolimerModelConfig<TModel extends ImmutableModel> {
     stateHandlerModelsConfig: Map<string, StateHandlerModelMetadata>;
     stateHandlersConfig: Map<string, StateHandlerConfiguration[]>;
     eventTransformConfig: EventTransformConfiguration[];
+    devToolsConfig: DevToolsConfig<TModel>;
 }
 
 export interface PolimerModelConfigUpdate {
@@ -54,11 +73,12 @@ export class PolimerModel<TModel extends ImmutableModel> extends DisposableBase 
     private readonly _modelEventHandlersByEventName: Map<string, ModelHandlerMetadata<TModel>[]> = new Map();
     private _disposablesKeyedOnObjectScannedAtWireup: Map<object, Disposable> = new Map();
     private _activeStateHandlerModels: { stateName: keyof TModel; model: StateHandlerModel<any> }[] = [];
-    private _immutableModel: TModel;
+    private _immutableModelUtility: ImmutableModelUtility<TModel>;
     private _modelPreEventProcessor: ModelPreEventProcessor<TModel>;
     private _modelPostEventProcessor: ModelPostEventProcessor<TModel>;
     private readonly _modelId: string;
     private _stateSaveHandler: (model: TModel) => any;
+    private _devToolsConfig: DevToolsConfig<TModel>;
 
     constructor(private readonly _router: Router, initialConfig: PolimerModelConfig<TModel>) {
         super();
@@ -67,7 +87,7 @@ export class PolimerModel<TModel extends ImmutableModel> extends DisposableBase 
         Guard.isObject(initialConfig.initialModel, 'initialConfig.initialModel must be defined');
         Guard.stringIsNotEmpty(initialConfig.initialModel.modelId, `initialConfig.initialModel.modelId must not be null or empty`);
         this._modelId = initialConfig.initialModel.modelId;
-        this._immutableModel = initialConfig.initialModel;
+        this._immutableModelUtility = createImmutableModelUtility(this.modelId, initialConfig.initialModel);
         this._stateSaveHandler = initialConfig.stateSaveHandler;
         if (initialConfig.modelPreEventProcessor) {
             Guard.isFunction(initialConfig.modelPreEventProcessor, 'The initialConfig.modelPreEventProcessor is not a function');
@@ -78,12 +98,11 @@ export class PolimerModel<TModel extends ImmutableModel> extends DisposableBase 
             this._modelPostEventProcessor = initialConfig.modelPostEventProcessor;
         }
         this._router.addModel(this._modelId, this);
-        connectDevTools(this._router, this._modelId, this, this._modelId);
-        sendUpdateToDevTools('@@INIT', this._immutableModel, this._modelId);
         this._wireUpStateHandlerModels(initialConfig.stateHandlerModelsConfig);
         this._wireUpStateHandlers(initialConfig.stateHandlersConfig);
         this._wireUpEventTransforms(initialConfig.eventTransformConfig);
         this.addDisposable(this._router.observeEventsOn(this._modelId, this));
+        this._setDevToolsConfig(initialConfig.devToolsConfig);
     }
 
     public get modelId() {
@@ -121,28 +140,30 @@ export class PolimerModel<TModel extends ImmutableModel> extends DisposableBase 
     };
 
     preProcess() {
+        this._immutableModelUtility.beginMutation();
         if (this._modelPreEventProcessor) {
-            const newModel = this._modelPreEventProcessor(this._immutableModel);
-            // has the model been replaced by the processor?
+            const newModel = this._modelPreEventProcessor(this._immutableModelUtility.model);
+            // has the entire model been replaced by the processor?
             if (newModel) {
-                this._immutableModel = newModel;
+                this._immutableModelUtility.replaceModel(newModel);
             }
         }
+
         // run pre-processing for the OO/legacy type models which may be integrating with polimer models
         this._activeStateHandlerModels.forEach(({model, stateName}) => {
             if (model.preProcess) {
                 model.preProcess(model);
-                this._immutableModel[stateName] = model.getEspPolimerState();
+                this._immutableModelUtility.model[stateName] = model.getEspPolimerState();
             }
         });
     }
 
     postProcess(eventsProcessed: string[]) {
         if (this._modelPostEventProcessor) {
-            const newModel = this._modelPostEventProcessor(this._immutableModel, eventsProcessed);
+            const newModel = this._modelPostEventProcessor(this._immutableModelUtility.model, eventsProcessed);
             // has the model been replaced by the processor?
             if (newModel) {
-                this._immutableModel = newModel;
+                this._immutableModelUtility.replaceModel(newModel);
             }
         }
         // run post-processing for the OO/legacy type models which may be integrating with polimer models
@@ -150,9 +171,10 @@ export class PolimerModel<TModel extends ImmutableModel> extends DisposableBase 
             if (model.postProcess) {
                 model.postProcess(model, eventsProcessed);
                 const nextState = model.getEspPolimerState();
-                this._immutableModel[stateName] = nextState;
+                this._immutableModelUtility.model[stateName] = nextState;
             }
         });
+        this._immutableModelUtility.endMutation();
     }
 
     /**
@@ -160,18 +182,35 @@ export class PolimerModel<TModel extends ImmutableModel> extends DisposableBase 
      * Polimer doesn't have a hard dependency on esp-js-ui, however if your models are created using esp-js-ui then this hook will be used to save state.
      *
      * */
-    getEspUiModelState(): any {
+    public getEspUiModelState = (): any => {
         if (!this._stateSaveHandler) {
             return null;
         }
-        return this._stateSaveHandler(this._immutableModel);
+        return this._stateSaveHandler(this._immutableModelUtility.model);
+    };
+
+    /**
+     * @deprecated
+     * Previously this returns a non changing instance, hence it now points to getEspPolimerImmutableModelProxy().
+     */
+    public getImmutableModel = (): TModel => {
+        return this.getEspPolimerImmutableModelProxy();
+    };
+
+    /**
+     * Returns a Proxy which always backs onto the latest PolimerModel.getEspPolimerImmutableModel();
+     * The returned Proxy instance won't chant.
+     */
+    getEspPolimerImmutableModelProxy() {
+        return this._immutableModelUtility.modelProxy;
     }
 
     /**
-     * A convention-named function used by esp-js-react to select the model to pass to a view connected via ConnectableComponent.
+     * Returns the latest model state.
+     * This instance will change when the esp Router has finished processing events against this model.
      */
-    getEspReactRenderModel() {
-        return this.getImmutableModel();
+    getEspPolimerImmutableModel() {
+        return this._immutableModelUtility.model;
     }
 
     // called by the router when it's finished dispatching an event
@@ -183,10 +222,10 @@ export class PolimerModel<TModel extends ImmutableModel> extends DisposableBase 
         if (handlers) {
             handlers.forEach((modelHandlerMetadata: ModelHandlerMetadata<TModel>) => {
                 // Given an event processed by the model in question has just finished, we replace the relevant state on the immutable model
-                (<any>this._immutableModel[modelHandlerMetadata.stateName]) = modelHandlerMetadata.model.getEspPolimerState();
+                (<any>this._immutableModelUtility.model[modelHandlerMetadata.stateName]) = modelHandlerMetadata.model.getEspPolimerState();
             });
-            sendUpdateToDevTools({eventType: eventType, event: event}, this._immutableModel, this._modelId);
         }
+        this._trySendDevToolsUpdate(eventType, event);
     }
 
     @observeEvent(PolimerEvents.disposeModel)
@@ -262,7 +301,7 @@ export class PolimerModel<TModel extends ImmutableModel> extends DisposableBase 
                             .getEventObservable(this._modelId, decoratorMetadata.eventType, decoratorMetadata.observationStage)
                             .filter(eventEnvelope => eventIsForThisSpecificHandlerPredicate(eventEnvelope))
                             .subscribe((eventEnvelope: EventEnvelope<any, any>) => {
-                                const model = <any>this._immutableModel;
+                                const model = <any>this._immutableModelUtility.model;
                                 let stateReaderWriter: StateReaderWriter = getStateReaderWriter(eventEnvelope.entityKey);
                                 const beforeState = stateReaderWriter.getState(eventEnvelope.entityKey);
                                 let processEvent = true;
@@ -419,22 +458,14 @@ export class PolimerModel<TModel extends ImmutableModel> extends DisposableBase 
     };
 
     private _mapEventEnvelopToInputEvent(eventEnvelope: EventEnvelope<any, PolimerModel<TModel>>): InputEvent<TModel, any> {
-        const immutableModel: TModel = eventEnvelope.model.getImmutableModel();
+        const modelProxy: TModel = eventEnvelope.model.getEspPolimerImmutableModelProxy();
         return <InputEvent<TModel, any>>{
             event: eventEnvelope.event,
             eventType: eventEnvelope.eventType,
-            model: immutableModel,
+            model: modelProxy,
             context: eventEnvelope.context
         };
     }
-
-    public getImmutableModel = (): TModel => {
-        return this._immutableModel;
-    };
-
-    public setImmutableModel = (value: TModel) => {
-        this._immutableModel = value;
-    };
 
     private _getDisposableForObject = (trackedObject: any) => {
         Guard.isFalsey(this._disposablesKeyedOnObjectScannedAtWireup.has(trackedObject), 'Object already registered. Using the same object instance for multiple esp-js-polimer registrations (be it state handlers or event transforms) is not supported.');
@@ -452,21 +483,55 @@ export class PolimerModel<TModel extends ImmutableModel> extends DisposableBase 
         // While we can know at event subscription if the state on the model is a Map, we don't know if an event will want to target an item in the map, or the entire map.
         // Which of the two will be known if the event is published with an 'entityKey', which is an addressing property the esp router supports.
         // At this point, we can create an API which works with both cases.
-        const stateTypeIsMap = this._immutableModel[stateName] instanceof Map;
+        const stateTypeIsMap = this._immutableModelUtility.model[stateName] instanceof Map;
         const mapReaderWriter: StateReaderWriter = stateTypeIsMap
-            ? new MapReaderWriter(() => this._immutableModel, stateName)
+            ? new MapReaderWriter(() => this._immutableModelUtility.model, stateName)
             : undefined;
-        const directStateReaderWriter: StateReaderWriter = new DirectStateReaderWriter(() => this._immutableModel, stateName);
+        const directStateReaderWriter: StateReaderWriter = new DirectStateReaderWriter(() => this._immutableModelUtility.model, stateName);
         return (entityKey: string) => {
             return stateTypeIsMap && utils.isString(entityKey)
                 ? mapReaderWriter
                 : directStateReaderWriter;
         };
     };
+
+    private _setDevToolsConfig(devToolsConfig: DevToolsConfig<TModel>) {
+        if (!devToolsConfig) {
+            return;
+        }
+        Guard.isFunction(devToolsConfig.devToolsStateSelector, `DevTools enabled for ${this._modelId}, however [${devToolsConfig.devToolsStateSelector}] is not set or not a function`);
+        this._devToolsConfig = {
+            ignoredEvents: devToolsConfig.ignoredEvents || [],
+            devToolsStateSelector: devToolsConfig.devToolsStateSelector,
+        };
+        const unsubscribe = connectReduxDevTools(this._modelId);
+        this.addDisposable(() => {
+            unsubscribe();
+        });
+        this._trySendDevToolsUpdate('@@INIT', null);
+    }
+
+    private _trySendDevToolsUpdate = (eventType: string, event: any) => {
+        if (!this._devToolsConfig) {
+            return;
+        }
+        if (!reduxDevToolsDetectedAndEnabledInEsp()) {
+            return;
+        }
+        if (this._devToolsConfig.ignoredEvents.includes(eventType)) {
+            return;
+        }
+        let state = this._devToolsConfig.devToolsStateSelector(this._immutableModelUtility.model);
+        sendUpdateToReduxDevTools(
+            {eventType: eventType, event: event},
+            state,
+            this._modelId
+        );
+    };
 }
 
 export namespace PolimerModel {
     export const isPolimerModel = (obj: any): obj is PolimerModel<any> => {
-        return 'getImmutableModel' in obj;
+        return 'getEspPolimerImmutableModel' in obj;
     };
 }
