@@ -18,41 +18,51 @@
 //
 
 // import for side effects
-import './polimer/polimerExtentsions';
+import './polimer/polimerModelBuilderExtentsions';
 
-export {RouterProvider, RouterContext, useRouter} from './routerProvider';
-export {SmartComponent} from './smartComponent';
 export {ViewBinder} from './viewBinder';
 export {viewBinding, DEFAULT_VIEW_KEY} from './viewBindingDecorator';
-export * from './getEspReactRenderModel';
-export {shouldUpdateMixin} from './shouldUpdateMixin';
 export {createViewForModel} from './viewBindingDecorator';
 export {
-    ConnectableComponent,
     connect,
-    ConnectableComponentProps,
-    MapModelToProps,
+    ConnectableComponentFactory
+} from './connect';
+export {
     CreatePublishEventProps,
-    ConnectableView,
+    MapModelToProps,
+    ConnectableComponentProps,
     ConnectableComponentChildProps,
+    ConnectableComponent,
 } from './connectableComponent';
 export {
-    useGetModelId,
-    GetModelIdDelegate,
-    GetModelIdContext,
-
+    useSyncModelWithSelector,
+    SyncModelWithSelectorOptionsBuilder,
+    syncModelWithSelectorOptions,
+    SyncModelWithSelectorOptions,
+    SyncModelWithSelectorEqualityFn,
+} from './useSyncModelWithSelector';
+export {
+    RouterProvider,
+    RouterProviderProps,
+    EspRouterContextProvider,
+    EspRouterContextProviderProps,
+    RouterContext,
+    useRouter,
     PublishEventDelegate,
     PublishEventContext,
     usePublishEvent,
-
+} from './espRouterContextProvider';
+export {
+    useGetModelId,
+    GetModelIdContext,
+    useGetModel,
+    GetModelContext,
     PublishModelEventDelegate,
     PublishModelEventContext,
     usePublishModelEvent,
-
     PublishModelEventWithEntityKeyDelegate,
     PublishModelEventWithEntityKeyContext,
     usePublishModelEventWithEntityKey,
-
-    EspModelContext,
-    EspModelContextProps
-} from './espModelContext';
+    EspModelContextProvider,
+    EspModelContextProviderProps,
+} from './espModelContextProvider';
