@@ -1,6 +1,5 @@
 // Auto-generated 
-export * from './compositeDiagnosticMonitor';
-export * from './devToolsDiagnosticMonitor';
 export * from './diagnosticMonitor';
-export * from './loggingDiagnosticMonitor';
 export * from './noopDiagnosticMonitor';
+export * from './reduxDevToolsConnector';
+export * from './reduxDevToolsDiagnosticMonitor';
